requirements.txt
streamlit
openai
python-dotenv
elevenlabs
streamlit-audiorecorder
streamlit-webrtc
SpeechRecognition
pydub
requests
streamlit-audiorec